refactor: meeting 엔티티 패키지 변경
