+++ src/main/java/com/mobble/mobbleserver/domain/meeting/Meeting.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.infrastructure.web.meeting;
+package com.mobble.mobbleserver.domain.meeting;
 
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.meeting.MeetingErrorCode;
